+++ src/renderer/features/governance/model/unlock/unlock.ts
@@ -1,27 +1,17 @@
 import { type ApiPromise } from '@polkadot/api';
 import { type BN, BN_ZERO } from '@polkadot/util';
-import { createEffect, createEvent, createStore, restore, sample } from 'effector';
-import { combineEvents, spread } from 'patronum';
+import { createEffect, createStore, sample } from 'effector';
+import { combineEvents } from 'patronum';
 
 import { type ClaimTimeAt, type UnlockChunk, UnlockChunkType } from '@shared/api/governance';
 import { type Address, type Referendum, type TrackId, type TrackInfo, type VotingMap } from '@shared/core';
-import { Step, getCreatedDateFromApi, getCurrentBlockNumber } from '@shared/lib/utils';
-import { claimScheduleService, referendumModel } from '@entities/governance';
+import { getCreatedDateFromApi, getCurrentBlockNumber } from '@shared/lib/utils';
+import { claimScheduleService, referendumModel, tracksModel, votingModel } from '@entities/governance';
 import { walletModel } from '@entities/wallet';
-import { tracksAggregate } from '../../aggregates/tracks';
-import { confirmUnlockAggregate } from '../../aggregates/unlock';
-import { votingAggregate } from '../../aggregates/voting';
 import { unlockService } from '../../lib/unlockService';
 import { locksModel } from '../locks';
 import { networkSelectorModel } from '../networkSelector';
 
-const flowStarted = createEvent();
-const flowFinished = createEvent();
-const stepChanged = createEvent<Step>();
-const unlockConfirm = createEvent();
-const txSaved = createEvent();
-
-const $step = restore<Step>(stepChanged, Step.NONE);
 const $claimSchedule = createStore<UnlockChunk[]>([]).reset(walletModel.$activeWallet);
 const $totalUnlock = createStore<BN>(BN_ZERO);
 
@@ -69,13 +59,13 @@ const getClaimScheduleFx = createEffect(
 sample({
   clock: [
     referendumModel.events.requestDone,
-    combineEvents([locksModel.events.requestDone, votingAggregate.events.requestDone]),
+    combineEvents([locksModel.events.requestDone, votingModel.effects.requestVotingFx.done]),
   ],
   source: {
     api: networkSelectorModel.$governanceChainApi,
-    tracks: tracksAggregate.$tracks,
+    tracks: tracksModel.$tracks,
     trackLocks: locksModel.$trackLocks,
-    voting: votingAggregate.$voting,
+    voting: votingModel.$voting,
     referendums: referendumModel.$referendums,
     chain: networkSelectorModel.$governanceChain,
   },
@@ -109,62 +99,8 @@ sample({
   target: $totalUnlock,
 });
 
-sample({
-  clock: flowStarted,
-  fn: () => Step.INIT,
-  target: stepChanged,
-});
-
-sample({
-  clock: txSaved,
-  fn: () => Step.BASKET,
-  target: stepChanged,
-});
-
-sample({
-  clock: stepChanged,
-  target: $step,
-});
-
-sample({
-  clock: flowFinished,
-  fn: () => Step.NONE,
-  target: stepChanged,
-});
-
-sample({
-  clock: unlockConfirm,
-  source: {
-    chain: networkSelectorModel.$governanceChain,
-    asset: locksModel.$asset,
-    unlockableClaims: $claimSchedule.map((c) => c.filter((claim) => claim.type === UnlockChunkType.CLAIMABLE)),
-    totalUnlock: $totalUnlock,
-  },
-  filter: ({ chain, asset, totalUnlock }) => !!chain && !!asset,
-  fn: ({ chain, unlockableClaims, asset, totalUnlock }) => ({
-    event: { chain: chain!, unlockableClaims, asset: asset!, amount: totalUnlock.toString() },
-    step: Step.CONFIRM,
-  }),
-  target: spread({
-    event: confirmUnlockAggregate.events.formInitiated,
-    step: stepChanged,
-  }),
-});
-
 export const unlockModel = {
-  $step,
-  $pendingSchedule: $claimSchedule.map((c) => c.filter((claim) => claim.type !== UnlockChunkType.CLAIMABLE)),
-  $isLoading: getClaimScheduleFx.pending || referendumModel.$isReferendumsLoading,
-  $isUnlockable: $claimSchedule.map((c) => c.some((claim) => claim.type === UnlockChunkType.CLAIMABLE)),
+  $isLoading: getClaimScheduleFx.pending,
   $totalUnlock,
-
-  events: {
-    flowStarted,
-    stepChanged,
-    unlockConfirm,
-    txSaved,
-  },
-  output: {
-    flowFinished,
-  },
+  $claimSchedule,
 };
